# docs/swagger.yaml
basePath: /api/
definitions:
  gainprojection.ResponseDefault:
    type: object
  service.CategoryResponse:
    properties:
      category:
        type: string
      id:
        type: integer
    type: object
  service.CreateGainRequest:
    properties:
      pay_in:
        type: string
      value:
        type: number
    type: object
  service.CreateRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      is_passive:
        type: boolean
      pay_in:
        type: string
      recurrence:
        type: integer
      value:
        type: number
    type: object
  service.GainProjectionPaginateResponse:
    properties:
      current_page:
        type: integer
      page_limit:
        type: integer
      records:
        items:
          $ref: '#/definitions/service.GainProjectionResponse'
        type: array
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  service.GainProjectionResponse:
    properties:
      category:
        $ref: '#/definitions/service.CategoryResponse'
      description:
        type: string
      id:
        type: string
      is_passive:
        type: boolean
      pay_in:
        type: string
      recurrence:
        type: integer
      value:
        type: number
    type: object
  service.GainResponse:
    properties:
      category:
        $ref: '#/definitions/service.CategoryResponse'
      description:
        type: string
      gain_projection_id:
        type: string
      id:
        type: string
      is_passive:
        type: boolean
      pay_in:
        type: string
      value:
        type: number
    type: object
  service.UpdateRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      is_passive:
        type: boolean
      pay_in:
        type: string
      value:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: API que dispões de recursos para gerenciar as finanças pessoais
  title: Wallet Core
  version: 0.1.0
paths:
  /v1/gain-projection:
    get:
      consumes:
      - application/json
      description: Este endpoint permite obter uma listagem de receitas previstas
      parameters:
      - description: O número de registros retornados pela busca
        in: query
        name: page_size
        type: string
      - description: A página que será buscada
        in: query
        name: page
        type: string
      - description: O mês que será filtrado a busca
        in: query
        name: month
        required: true
        type: string
      - description: O ano que será filtrado a busca
        in: query
        name: year
        required: true
        type: string
      - description: Token de autenticação do usuário
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: Informações do usuário em base64
        in: header
        name: X-Userinfo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.GainProjectionPaginateResponse'
      summary: Obter uma listagem de Receitas Previstas
      tags:
      - Gain-Projection
    post:
      consumes:
      - application/json
      description: Este endpoint permite criar uma receita prevista
      parameters:
      - description: Modelo de criação da receita prevista
        in: body
        name: gain_projection
        required: true
        schema:
          $ref: '#/definitions/service.CreateRequest'
      - description: Token de autenticação do usuário
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: Informações do usuário em base64
        in: header
        name: X-Userinfo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.GainProjectionResponse'
      summary: Criar uma Receita Prevista
      tags:
      - Gain-Projection
  /v1/gain-projection/{id}:
    delete:
      consumes:
      - application/json
      description: Este endpoint permite remover uma receita prevista
      parameters:
      - description: Id da receita prevista
        in: path
        name: id
        required: true
        type: string
      - description: Token de autenticação do usuário
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: Informações do usuário em base64
        in: header
        name: X-Userinfo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/gainprojection.ResponseDefault'
            - properties:
                message:
                  type: string
                status:
                  type: integer
              type: object
      summary: Remove uma Receita Prevista
      tags:
      - Gain-Projection
    get:
      consumes:
      - application/json
      description: Este endpoint permite obter uma receita prevista
      parameters:
      - description: Id da receita prevista
        in: path
        name: id
        required: true
        type: string
      - description: Token de autenticação do usuário
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: Informações do usuário em base64
        in: header
        name: X-Userinfo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.GainProjectionResponse'
      summary: Obter uma Receita Prevista
      tags:
      - Gain-Projection
    put:
      consumes:
      - application/json
      description: Este endpoint permite editar uma receita prevista
      parameters:
      - description: Modelo de edição da receita prevista
        in: body
        name: gain_projection
        required: true
        schema:
          $ref: '#/definitions/service.UpdateRequest'
      - description: Token de autenticação do usuário
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: Informações do usuário em base64
        in: header
        name: X-Userinfo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.GainProjectionResponse'
      summary: Editar uma Receita Prevista
      tags:
      - Gain-Projection
  /v1/gain-projection/{id}/create-gain:
    post:
      consumes:
      - application/json
      description: Este endpoint permite realizar uma receita que foi prevista
      parameters:
      - description: Id da receita prevista
        in: path
        name: id
        required: true
        type: string
      - description: Modelo de criação da receita
        in: body
        name: gain
        required: true
        schema:
          $ref: '#/definitions/service.CreateGainRequest'
      - description: Token de autenticação do usuário
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: Informações do usuário em base64
        in: header
        name: X-Userinfo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.GainResponse'
      summary: Realizar uma Receita Prevista
      tags:
      - Gain-Projection
swagger: "2.0"